WIP: documentation updates and moving stuff
