+++ .golangci.yaml
