+++ lib/doc.go
@@ -1,0 +1,8 @@
+// Copyright (c) 2024 Daniel Pivalizza. All rights reserved.
+// Use of this source code is governed by the MIT License that can be found in
+// the LICENSE file.
+
+// Package environ facilitates loading values for a struct by tags, where the tags
+// define the rules for a value and where it can be loaded from.
+// Currently supports environment variables.
+package environ
